pyproject.toml
[tool.poetry]
name = "image-analyze"
version = "0.1.0"
description = ""
authors = ["Narayan Bandodker <narayanband1356@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
openai = "^1.26.0"
pymupdf = "^1.24.2"
pillow = "^10.3.0"
python-datauri = "^2.1.0"
python-dotenv = "^1.0.1"


# Development dependencies

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
black = "^23.12.0"
pre-commit = "^3.7.0"
watchfiles = "^0.21.0"


# Tool configurations

## Black (formatter)

[tool.black]
line-length = 120
target-version = ['py39']
include = '\.pyi?$'


## Flake8 (PEP-8 linter)

[tool.flake8]
ignore = "E203, E266, E501, W503, F403, F401"
exclude = ".git, __pycache__, venv, .vscode, .mypy_cache, .pytest_cache, build, dist"
max-line-length = 120
max-complexity = 18
select = "B,C,E,F,W,T4,B9"
accept-encodings = "utf-8"
inline-quotes = "single"
max-linenumber = 500
multiline-quotes = "double"


## Pyright (Static type checker)

[tool.pyright]
venvPath = "."
venv = "venv"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
